requirements.txt
absl-py==1.4.0
altgraph==0.17.3
annotated-types==0.5.0
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
cx-Freeze==6.15.6
cx-Logging==3.1.0
cycler==0.11.0
distlib==0.3.7
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.0
idna==3.4
joblib==1.3.2
kiwisolver==1.4.4
lief==0.13.2
load-dotenv==0.1.0
matplotlib==3.7.2
mediapipe==0.10.3
numpy==1.25.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.3
pefile==2023.2.7
Pillow==10.0.0
pipenv==2023.8.23
platformdirs==3.10.0
protobuf==3.20.3
pycparser==2.21
pydantic_core==2.6.3
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.2
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sounddevice==0.4.6
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.3
